[v0.89.1][maintainability] Split execute/state/runtime_control.rs by responsibility (Closes #1962)
